backend/requirements.txt
fastapi
uvicorn
openai
pinecone
python-dotenv
pydantic